+++ types/novnc.d.ts
@@ -8,6 +8,7 @@ declare module "@novnc/novnc/lib/rfb.js" {
     blur(): void;
     disconnect(): void;
     sendCtrlAltDel(): void;
+    clipboardPasteFrom(text: string): void;
     addEventListener(type: string, listener: (event: Event) => void): void;
     removeEventListener(type: string, listener: (event: Event) => void): void;
   }